ia-local/requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
flask-cors==4.0.0
PyPDF2==3.0.1
pdfplumber==0.10.3
python-multipart==0.0.6
reportlab==4.0.4 